refactor(api): share util
