+++ packages/taxonomy/src/index.ts
@@ -1,3 +1,4 @@
 export { TAXONOMY } from "./taxonomy";
 export type { EnrichableTaxonomyKey, GateTaxonomyKey, TaxonomyKey, TaxonomyListItem, TaxonomyValueItem, TaxonomyValueKey, ValueKey } from "./types";
-export { ENRICHABLE_TAXONOMIES, GATE_TAXONOMIES, getTaxonomyList, isValidTaxonomyValueKey } from "./utils";
+export { ENRICHABLE_TAXONOMIES, GATE_TAXONOMIES, getTaxonomyList, isValidTaxonomyValueKey, parseTaxonomyValueKey } from "./utils";
+export type { ParsedTaxonomyValueKey } from "./utils";
